refactor: Phase 4 decompose runner.ts into 6 modules, split step-registry.ts into 5 files
- Extract QueryExecutor, FlowOrchestrator, SchemaManager, CompletionManager, BoundaryHooks, ClosureAdapter from runner.ts
- runner.ts now slim coordinator delegating to extracted modules
- Split step-registry.ts (956 lines) into types/loader/validator/serializer/utils + mod.ts
- Both use re-export pattern for full backward compatibility
