+++ agents/runner/runner.ts
@@ -1,5 +1,5 @@
 /**
- * Agent Runner - main execution engine
+ * Agent Runner - main execution engine (slim coordinator)
  *
  * Dual-loop architecture:
  * - Flow Loop: Step advancement and handoff management
@@ -7,6 +7,14 @@
  *
  * Supports dependency injection for testability.
  * Use AgentRunnerBuilder for convenient construction with custom dependencies.
+ *
+ * Responsibilities are delegated to extracted modules:
+ * - QueryExecutor: SDK query execution, message processing, rate limiting
+ * - FlowOrchestrator: Step routing, transitions, stepId normalization
+ * - SchemaManager: Schema loading, fail-fast tracking, flow validation
+ * - CompletionManager: Completion initialization, validation, AI declaration detection
+ * - BoundaryHooks: Closure step boundary hook invocation
+ * - ClosureAdapter: Closure step prompt adaptation
  */
 
 import type {
@@ -15,74 +23,34 @@ import type {
   IterationSummary,
   RuntimeContext,
 } from "../src_common/types.ts";
-import { isRecord, isString } from "../src_common/type-guards.ts";
 import {
   AgentMaxIterationsError,
   AgentNotInitializedError,
-  AgentQueryError,
-  AgentRateLimitError,
-  AgentSchemaResolutionError,
   AgentStepRoutingError,
   isAgentError,
   normalizeToAgentError,
 } from "./errors.ts";
-import { calculateBackoff, isRateLimitError } from "./error-classifier.ts";
 import { getAgentDir } from "./loader.ts";
-import { mergeSandboxConfig, toSdkSandboxConfig } from "./sandbox-defaults.ts";
 import type { AgentDependencies } from "./builder.ts";
-import { createDefaultDependencies, isInitializable } from "./builder.ts";
-import type { CompletionValidator } from "../validators/completion/validator.ts";
-import type { RetryHandler } from "../retry/retry-handler.ts";
-import {
-  type ExtendedStepsRegistry,
-  hasCompletionChainSupport,
-  hasFlowRoutingSupport,
-  type OutputSchemaRef,
-} from "../common/completion-types.ts";
-import type {
-  PromptStepDefinition,
-  StepKind,
-} from "../common/step-registry.ts";
-import { inferStepKind, loadStepRegistry } from "../common/step-registry.ts";
-import {
-  filterAllowedTools,
-  getToolPolicy,
-  isBashCommandAllowed,
-} from "../common/tool-policy.ts";
-import {
-  CompletionChain,
-  type CompletionValidationResult,
-} from "./completion-chain.ts";
-import { join } from "@std/path";
-import { PATHS } from "../shared/paths.ts";
-import {
-  SchemaPointerError,
-  SchemaResolver,
-} from "../common/schema-resolver.ts";
-import {
-  isAssistantMessage,
-  isErrorMessage,
-  isResultMessage,
-  isToolUseMessage,
-} from "./message-types.ts";
+import { createDefaultDependencies } from "./builder.ts";
 import {
   type AgentEvent,
   AgentEventEmitter,
   type AgentEventHandler,
 } from "./events.ts";
-import { StepContextImpl } from "../loop/step-context.ts";
 import type { StepContext } from "../src_common/contracts.ts";
-import { StepGateInterpreter } from "./step-gate-interpreter.ts";
-import { type RoutingResult, WorkflowRouter } from "./workflow-router.ts";
-import type { StepRegistry } from "../common/step-registry.ts";
 import { createVerboseLogger, type VerboseLogger } from "./verbose-logger.ts";
 import { PromptLogger } from "../common/prompt-logger.ts";
-import {
-  createFallbackProvider,
-  PromptResolver as StepPromptResolver,
-} from "../common/prompt-resolver.ts";
-import { AGENT_LIMITS, TRUNCATION } from "../shared/constants.ts";
+import { AGENT_LIMITS } from "../shared/constants.ts";
 import { STEP_PHASE } from "../shared/step-phases.ts";
+
+// Extracted modules
+import { QueryExecutor } from "./query-executor.ts";
+import { FlowOrchestrator } from "./flow-orchestrator.ts";
+import { SchemaManager } from "./schema-manager.ts";
+import { CompletionManager } from "./completion-manager.ts";
+import { BoundaryHooks } from "./boundary-hooks.ts";
+import { ClosureAdapter } from "./closure-adapter.ts";
 
 export interface RunnerOptions {
   /** Working directory */
@@ -95,7 +63,6 @@ export interface RunnerOptions {
   verbose?: boolean;
 }
 
-// CompletionValidationResult is now imported from completion-chain.ts
 // Re-export for backward compatibility
 export type { CompletionValidationResult } from "./completion-chain.ts";
 
@@ -106,38 +73,19 @@ export class AgentRunner {
   private context: RuntimeContext | null = null;
   private args: Record<string, unknown> = {};
 
+  // Extracted module instances
+  private readonly completionManager: CompletionManager;
+  private readonly schemaManager: SchemaManager;
+  private readonly flowOrchestrator: FlowOrchestrator;
+  private readonly queryExecutor: QueryExecutor;
+  private readonly boundaryHooks: BoundaryHooks;
+  private readonly closureAdapter: ClosureAdapter;
+
   // Completion validation
-  private completionValidator: CompletionValidator | null = null;
-  private retryHandler: RetryHandler | null = null;
-  private stepsRegistry: ExtendedStepsRegistry | null = null;
-  private completionChain: CompletionChain | null = null;
   private pendingRetryPrompt: string | null = null;
-
-  // Rate limit handling
-  private rateLimitRetryCount = 0;
-  private static readonly MAX_RATE_LIMIT_RETRIES =
-    AGENT_LIMITS.MAX_RATE_LIMIT_RETRIES;
-
-  // Step flow orchestration
-  private stepContext: StepContextImpl | null = null;
-  private currentStepId: string | null = null;
-  private stepGateInterpreter: StepGateInterpreter | null = null;
-  private workflowRouter: WorkflowRouter | null = null;
-
-  // Schema resolution failure tracking (fail-fast)
-  // Maps stepId -> consecutive failure count
-  private schemaFailureCount: Map<string, number> = new Map();
-  // Flag to skip StepGate when schema resolution failed
-  private schemaResolutionFailed = false;
-  // Maximum consecutive schema failures before aborting
-  private static readonly MAX_SCHEMA_FAILURES =
-    AGENT_LIMITS.MAX_SCHEMA_FAILURES;
 
   // Verbose logging for SDK I/O debugging
   private verboseLogger: VerboseLogger | null = null;
-
-  // Step-level prompt resolver (new resolver for closure adaptation)
-  private stepPromptResolver: StepPromptResolver | null = null;
 
   /**
    * Create an AgentRunner with optional dependency injection.
@@ -146,6 +94,48 @@ export class AgentRunner {
     this.definition = definition;
     this.dependencies = dependencies ?? createDefaultDependencies();
     this.eventEmitter = new AgentEventEmitter();
+
+    // Initialize extracted modules with dependency bridges
+    this.completionManager = new CompletionManager({
+      definition: this.definition,
+      dependencies: this.dependencies,
+    });
+
+    this.schemaManager = new SchemaManager({
+      definition: this.definition,
+      getContext: () => this.getContext(),
+      getStepsRegistry: () => this.completionManager.stepsRegistry,
+    });
+
+    this.flowOrchestrator = new FlowOrchestrator({
+      definition: this.definition,
+      args: this.args,
+      getStepsRegistry: () => this.completionManager.stepsRegistry,
+      getStepGateInterpreter: () => this.completionManager.stepGateInterpreter,
+      getWorkflowRouter: () => this.completionManager.workflowRouter,
+      hasFlowRoutingEnabled: () =>
+        this.completionManager.hasFlowRoutingEnabled(),
+    });
+
+    this.queryExecutor = new QueryExecutor({
+      definition: this.definition,
+      getContext: () => this.getContext(),
+      getStepsRegistry: () => this.completionManager.stepsRegistry,
+      getVerboseLogger: () => this.verboseLogger,
+      getSchemaManager: () => this.schemaManager,
+    });
+
+    this.boundaryHooks = new BoundaryHooks({
+      getStepsRegistry: () => this.completionManager.stepsRegistry,
+      getEventEmitter: () => this.eventEmitter,
+    });
+
+    this.closureAdapter = new ClosureAdapter({
+      definition: this.definition,
+      args: this.args,
+      getStepPromptResolver: () => this.completionManager.stepPromptResolver,
+      getStepsRegistry: () => this.completionManager.stepsRegistry,
+    });
   }
 
   /**
@@ -207,7 +197,12 @@ export class AgentRunner {
     );
 
     // Initialize Completion validation if registry supports it
-    await this.initializeCompletionValidation(agentDir, cwd, logger);
+    await this.completionManager.initializeCompletionValidation(
+      agentDir,
+      cwd,
+      logger,
+      this.schemaManager,
+    );
 
     // Initialize verbose logger if enabled
     if (options.verbose) {
@@ -224,12 +219,10 @@ export class AgentRunner {
     const promptLogger = new PromptLogger(logger, {
       logSuccess: true,
       logFailures: true,
-      logVariables: true, // Log uv-* variables for usage analysis
+      logVariables: true,
     });
 
     // Inject prompt logger into resolver for automatic logging
-    // This enables logging of actual file paths (e.g., iterator/initial/issue/f_default.md)
-    // with edition, adaptation, and variable information
     promptResolver.setPromptLogger(promptLogger);
 
     // Assign all context at once (atomic initialization)
@@ -249,8 +242,7 @@ export class AgentRunner {
     const ctx = this.getContext();
 
     // Initialize step context for step flow orchestration
-    this.stepContext = new StepContextImpl();
-    this.currentStepId = this.getStepIdForIteration(1);
+    this.flowOrchestrator.initializeStepContext();
 
     // Emit initialized event
     await this.eventEmitter.emit("initialized", { cwd: ctx.cwd });
@@ -273,7 +265,7 @@ export class AgentRunner {
         ctx.logger.info(`=== Iteration ${iteration} ===`);
 
         // Verbose: Log iteration start
-        const stepId = this.getStepIdForIteration(iteration);
+        const stepId = this.flowOrchestrator.getStepIdForIteration(iteration);
         if (this.verboseLogger) {
           // deno-lint-ignore no-await-in-loop
           await this.verboseLogger.logIterationStart(iteration, stepId);
@@ -305,7 +297,8 @@ export class AgentRunner {
         } else {
           // Try closure adaptation for closure steps
           // deno-lint-ignore no-await-in-loop
-          const closurePrompt = await this.tryClosureAdaptation(stepId, ctx);
+          const closurePrompt = await this.closureAdapter
+            .tryClosureAdaptation(stepId, ctx);
           if (closurePrompt) {
             prompt = closurePrompt.content;
             promptSource = closurePrompt.source;
@@ -324,8 +317,6 @@ export class AgentRunner {
         const promptTimeMs = performance.now() - promptStartTime;
 
         // Log step prompt for usage analysis
-        // Note: Step prompts are constructed by CompletionHandler, not loaded from files
-        // The stepId identifies which step in the flow is being executed
         if (ctx.promptLogger) {
           // deno-lint-ignore no-await-in-loop
           await ctx.promptLogger.logResolution(
@@ -333,7 +324,6 @@ export class AgentRunner {
               stepId,
               source: promptSource,
               content: prompt,
-              // Include prompt type and completion type for analysis
               promptPath:
                 `${this.definition.behavior.completionType}/${promptType}`,
             },
@@ -341,7 +331,7 @@ export class AgentRunner {
           );
         }
 
-        // Resolve system prompt (logging is automatic via PromptResolver.setPromptLogger)
+        // Resolve system prompt
         // deno-lint-ignore no-await-in-loop
         const systemPromptResult = await ctx.promptResolver
           .resolveSystemPromptWithMetadata(
@@ -382,9 +372,9 @@ export class AgentRunner {
           await this.verboseLogger.logSystemPrompt(systemPrompt);
         }
 
-        // Execute Claude SDK query (stepId already determined at iteration start)
+        // Execute Claude SDK query
         // deno-lint-ignore no-await-in-loop
-        const summary = await this.executeQuery({
+        const summary = await this.queryExecutor.executeQuery({
           prompt,
           systemPrompt,
           plugins,
@@ -400,11 +390,20 @@ export class AgentRunner {
         summaries.push(summary);
         sessionId = summary.sessionId;
 
+        // Sync schema resolution state to flow orchestrator
+        this.flowOrchestrator.setSchemaResolutionFailed(
+          this.schemaManager.schemaResolutionFailed,
+        );
+
         // Normalize stepId in structured output (Flow owns canonical value)
-        this.normalizeStructuredOutputStepId(stepId, summary, ctx);
+        this.flowOrchestrator.normalizeStructuredOutputStepId(
+          stepId,
+          summary,
+          ctx,
+        );
 
         // Record step output for step flow orchestration
-        this.recordStepOutput(stepId, summary);
+        this.flowOrchestrator.recordStepOutput(stepId, summary, ctx);
 
         // Handle rate limit retry
         if (summary.rateLimitRetry) {
@@ -418,17 +417,19 @@ export class AgentRunner {
         }
 
         // Completion validation: trigger when AI declares complete via structured output
-        if (this.hasAICompletionDeclaration(summary)) {
-          const completionStepId = this.getCompletionStepId();
+        if (this.completionManager.hasAICompletionDeclaration(summary)) {
+          const completionStepId = this.completionManager
+            .getCompletionStepId();
           ctx.logger.info(
             "AI declared completion, validating external conditions",
           );
           // deno-lint-ignore no-await-in-loop
-          const validation = await this.validateCompletionConditions(
-            completionStepId,
-            summary,
-            ctx.logger,
-          );
+          const validation = await this.completionManager
+            .validateCompletionConditions(
+              completionStepId,
+              summary,
+              ctx.logger,
+            );
 
           if (!validation.valid) {
             this.pendingRetryPrompt = validation.retryPrompt ?? null;
@@ -438,18 +439,19 @@ export class AgentRunner {
           }
         }
 
-        // Step transition for step flow orchestration (moved before completion check)
-        // Structured gate routing can signal completion via signalCompletion
-        const routingResult = this.handleStepTransition(stepId, summary, ctx);
+        // Step transition for step flow orchestration
+        const routingResult = this.flowOrchestrator.handleStepTransition(
+          stepId,
+          summary,
+          ctx,
+        );
 
         // R4: Fail-fast if iteration > 1 and no intent produced (no routing)
-        // This prevents silent rerun of entry step when StepGate can't parse intent
-        // @see agents/docs/design/08_step_flow_design.md Section 6
         if (
           iteration > 1 &&
           routingResult === null &&
           !summary.schemaResolutionFailed &&
-          this.hasFlowRoutingEnabled()
+          this.completionManager.hasFlowRoutingEnabled()
         ) {
           const errorMsg =
             `[StepFlow] No intent produced for iteration ${iteration} on step "${stepId}". ` +
@@ -481,9 +483,8 @@ export class AgentRunner {
           );
 
           // Invoke boundary hook for closure steps
-          // @see agents/docs/design/08_step_flow_design.md Section 7.1
           // deno-lint-ignore no-await-in-loop
-          await this.invokeBoundaryHook(stepId, summary, ctx);
+          await this.boundaryHooks.invokeBoundaryHook(stepId, summary, ctx);
         } else {
           // Fall back to legacy completionHandler
           if (ctx.completionHandler.setCurrentSummary) {
@@ -626,284 +627,11 @@ export class AgentRunner {
     }
   }
 
-  private async executeQuery(options: {
-    prompt: string;
-    systemPrompt: string | {
-      type: "preset";
-      preset: "claude_code";
-      append?: string;
-    };
-    plugins: string[];
-    sessionId?: string;
-    iteration: number;
-    stepId?: string;
-  }): Promise<IterationSummary> {
-    const { prompt, systemPrompt, plugins, sessionId, iteration, stepId } =
-      options;
-    const ctx = this.getContext();
-
-    const summary: IterationSummary = {
-      iteration,
-      sessionId: undefined,
-      assistantResponses: [],
-      toolsUsed: [],
-      errors: [],
-    };
-
-    try {
-      // Dynamic import of Claude Code SDK
-      const { query } = await import("@anthropic-ai/claude-agent-sdk");
-
-      // Apply stepKind-based tool gating if we have step info
-      let allowedTools = this.definition.behavior.allowedTools;
-      let currentStepKind: StepKind | undefined;
-
-      if (stepId && this.stepsRegistry) {
-        const stepDef = this.stepsRegistry.steps[stepId] as
-          | PromptStepDefinition
-          | undefined;
-        if (stepDef) {
-          currentStepKind = inferStepKind(stepDef);
-          if (currentStepKind) {
-            allowedTools = filterAllowedTools(allowedTools, currentStepKind);
-            ctx.logger.info(
-              `[ToolPolicy] Step "${stepId}" (${currentStepKind}): tools filtered to ${allowedTools.length} allowed`,
-            );
-          }
-        }
-      }
-
-      const queryOptions: Record<string, unknown> = {
-        cwd: ctx.cwd,
-        systemPrompt,
-        allowedTools,
-        permissionMode: this.definition.behavior.permissionMode,
-        settingSources: ["user", "project"],
-        plugins,
-        resume: sessionId,
-        // Auto-respond to AskUserQuestion to enable autonomous execution
-        // Instead of waiting for user input, delegate decision to Claude
-        canUseTool: (
-          toolName: string,
-          input: Record<string, unknown>,
-        ) => {
-          if (toolName === "AskUserQuestion") {
-            const autoResponse = this.definition.behavior.askUserAutoResponse ??
-              "Use your best judgment to choose the optimal approach. No need to confirm again.";
-            const questions = input.questions as Array<{
-              question: string;
-              options: Array<{ label: string }>;
-            }>;
-            const answers: Record<string, string> = {};
-            for (const q of questions) {
-              // Delegate decision to Claude using configured response
-              answers[q.question] = autoResponse;
-            }
-            ctx.logger.info(
-              "[AskUserQuestion] Auto-responding with delegation",
-              { questionCount: questions.length, response: autoResponse },
-            );
-            return {
-              behavior: "allow",
-              updatedInput: { questions: input.questions, answers },
-            };
-          }
-          // Allow other tools
-          return { behavior: "allow", updatedInput: input };
-        },
-      };
-
-      // Configure sandbox
-      const sandboxConfig = mergeSandboxConfig(
-        this.definition.behavior.sandboxConfig,
-      );
-      if (sandboxConfig.enabled === false) {
-        queryOptions.dangerouslySkipPermissions = true;
-      } else {
-        queryOptions.sandbox = toSdkSandboxConfig(sandboxConfig);
-      }
-
-      // Configure structured output if step has outputSchemaRef
-      if (stepId) {
-        const schema = await this.loadSchemaForStep(
-          stepId,
-          iteration,
-          ctx.logger,
-        );
-
-        // R2: If schema resolution failed, abort iteration immediately
-        // Don't let LLM produce freeform text when schemas are unavailable
-        if (this.schemaResolutionFailed) {
-          ctx.logger.warn(
-            `[StructuredOutput] Aborting iteration: schema resolution failed for step "${stepId}"`,
-          );
-          summary.errors.push(
-            `Schema resolution failed for step "${stepId}". Iteration aborted.`,
-          );
-          summary.schemaResolutionFailed = true;
-          return summary;
-        }
-
-        if (schema) {
-          queryOptions.outputFormat = {
-            type: "json_schema",
-            schema,
-          };
-          ctx.logger.info(
-            `[StructuredOutput] Using schema for step: ${stepId}`,
-          );
-        }
-      }
-
-      // Configure PreToolUse hooks for boundary bash blocking
-      // Enable based on tool policy's blockBoundaryBash setting for each stepKind
-      if (currentStepKind && getToolPolicy(currentStepKind).blockBoundaryBash) {
-        const boundaryBashBlockingHook = this.createBoundaryBashBlockingHook(
-          currentStepKind,
-          ctx,
-        );
-        queryOptions.hooks = {
-          PreToolUse: [
-            {
-              matcher: "Bash",
-              hooks: [boundaryBashBlockingHook],
-            },
-          ],
-        };
-        ctx.logger.info(
-          `[ToolPolicy] PreToolUse hooks enabled for boundary bash blocking (stepKind: ${currentStepKind})`,
-        );
-      }
-
-      // Verbose: Log full SDK request options
-      if (this.verboseLogger) {
-        await this.verboseLogger.logSdkRequest({
-          ...queryOptions,
-          // Exclude canUseTool callback (not serializable)
-          canUseTool: "[Function]",
-        });
-      }
-
-      const queryIterator = query({
-        prompt,
-        options: queryOptions,
-      });
-
-      for await (const message of queryIterator) {
-        ctx.logger.logSdkMessage(message);
-
-        // Verbose: Log raw SDK message
-        if (this.verboseLogger) {
-          await this.verboseLogger.logSdkMessage(message);
-        }
-
-        this.processMessage(message, summary);
-      }
-    } catch (error) {
-      const errorMessage = error instanceof Error
-        ? error.message
-        : String(error);
-
-      // Check for rate limit error
-      if (isRateLimitError(errorMessage)) {
-        this.rateLimitRetryCount++;
-
-        if (this.rateLimitRetryCount >= AgentRunner.MAX_RATE_LIMIT_RETRIES) {
-          const rateLimitError = new AgentRateLimitError(
-            `Rate limit exceeded after ${this.rateLimitRetryCount} retries`,
-            {
-              attempts: this.rateLimitRetryCount,
-              cause: error instanceof Error ? error : undefined,
-              iteration,
-            },
-          );
-          summary.errors.push(rateLimitError.message);
-          ctx.logger.error("Rate limit retries exhausted", {
-            error: rateLimitError.message,
-            attempts: this.rateLimitRetryCount,
-          });
-          throw rateLimitError;
-        }
-
-        const waitTime = calculateBackoff(this.rateLimitRetryCount - 1);
-        ctx.logger.warn(
-          `Rate limit hit, waiting ${waitTime}ms before retry ` +
-            `(attempt ${this.rateLimitRetryCount}/${AgentRunner.MAX_RATE_LIMIT_RETRIES})`,
-        );
-
-        summary.errors.push(`Rate limit hit, will retry after ${waitTime}ms`);
-        summary.rateLimitRetry = {
-          waitMs: waitTime,
-          attempt: this.rateLimitRetryCount,
-        };
-      } else {
-        // Non-rate-limit error: reset counter
-        this.rateLimitRetryCount = 0;
-
-        const queryError = new AgentQueryError(errorMessage, {
-          cause: error instanceof Error ? error : undefined,
-          iteration,
-        });
-        summary.errors.push(queryError.message);
-        ctx.logger.error("Query execution failed", {
-          error: queryError.message,
-          code: queryError.code,
-          iteration: queryError.iteration,
-        });
-      }
-    }
-
-    return summary;
-  }
-
-  private processMessage(message: unknown, summary: IterationSummary): void {
-    const ctx = this.getContext();
-
-    if (isAssistantMessage(message)) {
-      const content = this.extractContent(message.message);
-      if (content) {
-        summary.assistantResponses.push(content);
-      }
-    } else if (isToolUseMessage(message)) {
-      summary.toolsUsed.push(message.tool_name);
-    } else if (isResultMessage(message)) {
-      summary.sessionId = message.session_id;
-      if (message.structured_output) {
-        summary.structuredOutput = message.structured_output;
-        ctx.logger.info("[StructuredOutput] Got structured output from result");
-      }
-      if (message.total_cost_usd !== undefined) {
-        summary.totalCostUsd = message.total_cost_usd;
-      }
-      if (message.num_turns !== undefined) {
-        summary.numTurns = message.num_turns;
-      }
-      if (message.duration_ms !== undefined) {
-        summary.durationMs = message.duration_ms;
-      }
-    } else if (isErrorMessage(message)) {
-      summary.errors.push(message.error.message ?? "Unknown error");
-    }
-  }
-
-  private extractContent(message: unknown): string {
-    if (isString(message)) {
-      return message;
-    }
-    if (isRecord(message)) {
-      if (isString(message.content)) {
-        return message.content;
-      }
-      if (Array.isArray(message.content)) {
-        return message.content
-          .filter((c): c is Record<string, unknown> =>
-            isRecord(c) && c.type === "text"
-          )
-          .map((c) => isString(c.text) ? c.text : "")
-          .join("\n");
-      }
-    }
-    return "";
+  /**
+   * Get the step context for data passing between steps.
+   */
+  getStepContext(): StepContext | null {
+    return this.flowOrchestrator.getStepContext();
   }
 
   private getMaxIterations(): number {
@@ -920,939 +648,9 @@ export class AgentRunner {
   }
 
   /**
-   * Initialize Completion validation system
-   *
-   * Loads StepRegistry and initializes components based on registry capabilities:
-   * - CompletionChain support (completionPatterns/validators): Initialize CompletionValidator, RetryHandler, CompletionChain
-   * - Flow routing support (structuredGate in steps): Initialize StepGateInterpreter, WorkflowRouter
-   *
-   * A registry is always loaded if it exists and has either capability.
-   */
-  private async initializeCompletionValidation(
-    agentDir: string,
-    cwd: string,
-    logger: import("../src_common/logger.ts").Logger,
-  ): Promise<void> {
-    const registryPath = join(agentDir, PATHS.STEPS_REGISTRY);
-
-    try {
-      // Use loadStepRegistry for unified validation (fail-fast per design/08_step_flow_design.md)
-      // This validates: stepKind/allowedIntents consistency, entryStepMapping, intentSchemaRef format
-      //
-      // First load registry WITHOUT intent enum validation to get schemasBase
-      // Then validate enums with the correct schemasDir (honoring registry.schemasBase)
-      const registry = await loadStepRegistry(
-        this.definition.name,
-        "", // Not used when registryPath is provided
-        {
-          registryPath,
-          validateIntentEnums: false, // Defer enum validation
-        },
-      );
-      logger.debug(
-        "Registry validation passed (stepKind, entryStep, intentSchemaRef format)",
-      );
-
-      // Honor registry.schemasBase override per builder/01_quickstart.md
-      const schemasBase = registry.schemasBase ??
-        `.agent/${this.definition.name}/schemas`;
-      const schemasDir = join(cwd, schemasBase);
-
-      // Now validate intent schema enums with the correct schemasDir
-      const { validateIntentSchemaEnums } = await import(
-        "../common/step-registry.ts"
-      );
-      await validateIntentSchemaEnums(registry, schemasDir);
-      logger.debug(
-        "Intent schema enum validation passed",
-      );
-
-      // Check for extended registry capabilities
-      const hasCompletionChain = hasCompletionChainSupport(registry);
-      const hasFlowRouting = hasFlowRoutingSupport(registry);
-
-      // Fail-fast: stepMachine completion requires structuredGate on at least one step
-      // Per design/08_step_flow_design.md and builder/01_quickstart.md
-      if (
-        this.definition.behavior.completionType === "stepMachine" &&
-        !hasFlowRouting
-      ) {
-        throw new Error(
-          `[StepFlow][ConfigError] Agent "${this.definition.name}" uses completionType "stepMachine" ` +
-            `but registry has no steps with structuredGate. Add structuredGate to at least one step ` +
-            `or change completionType. See design/08_step_flow_design.md.`,
-        );
-      }
-
-      if (!hasCompletionChain && !hasFlowRouting) {
-        logger.debug(
-          "Registry has no extended capabilities (no completionPatterns, validators, or structuredGate), skipping setup",
-        );
-        return;
-      }
-
-      // Store registry with proper typing (use local variable for type narrowing)
-      const stepsRegistry: ExtendedStepsRegistry = registry;
-      this.stepsRegistry = stepsRegistry;
-
-      const capabilities: string[] = [];
-      if (hasCompletionChain) capabilities.push("CompletionChain");
-      if (hasFlowRouting) capabilities.push("FlowRouting");
-      logger.info(
-        `Loaded steps registry with capabilities: ${capabilities.join(", ")}`,
-      );
-
-      // Initialize CompletionChain components if supported
-      if (hasCompletionChain) {
-        // Initialize CompletionValidator factory
-        const validatorFactory = this.dependencies.completionValidatorFactory;
-        if (validatorFactory) {
-          if (isInitializable(validatorFactory)) {
-            await validatorFactory.initialize();
-          }
-          this.completionValidator = validatorFactory.create({
-            registry: stepsRegistry,
-            workingDir: cwd,
-            logger,
-            agentId: this.definition.name,
-          });
-          logger.debug("CompletionValidator initialized");
-        }
-
-        // Initialize RetryHandler factory
-        const retryFactory = this.dependencies.retryHandlerFactory;
-        if (retryFactory) {
-          if (isInitializable(retryFactory)) {
-            await retryFactory.initialize();
-          }
-          this.retryHandler = retryFactory.create({
-            registry: stepsRegistry,
-            workingDir: cwd,
-            logger,
-            agentId: this.definition.name,
-          });
-          logger.debug("RetryHandler initialized");
-        }
-
-        // Initialize CompletionChain
-        this.completionChain = new CompletionChain({
-          workingDir: cwd,
-          logger,
-          stepsRegistry: stepsRegistry,
-          completionValidator: this.completionValidator,
-          retryHandler: this.retryHandler,
-          agentId: this.definition.name,
-        });
-        logger.debug("CompletionChain initialized");
-      }
-
-      // Initialize Flow routing components if supported
-      if (hasFlowRouting) {
-        // Validate that all Flow steps have structuredGate and transitions
-        this.validateFlowSteps(stepsRegistry, logger);
-
-        this.stepGateInterpreter = new StepGateInterpreter();
-        this.workflowRouter = new WorkflowRouter(
-          stepsRegistry as unknown as StepRegistry,
-        );
-        logger.debug("StepGateInterpreter and WorkflowRouter initialized");
-
-        // Initialize step prompt resolver for closure adaptation
-        this.stepPromptResolver = new StepPromptResolver(
-          stepsRegistry as unknown as StepRegistry,
-          createFallbackProvider({}),
-          { workingDir: cwd, configSuffix: "steps" },
-        );
-        logger.debug("StepPromptResolver initialized for closure adaptation");
-      }
-    } catch (error) {
-      if (error instanceof Deno.errors.NotFound) {
-        logger.debug(
-          `Steps registry not found at ${registryPath}, using default completion`,
-        );
-      } else {
-        logger.warn(`Failed to load steps registry: ${error}`);
-      }
-    }
-  }
-
-  /**
-   * Validate that all Flow steps have structuredGate, transitions, and outputSchemaRef.
-   *
-   * Flow steps are all steps except those prefixed with "section." (template sections).
-   * Missing structuredGate, transitions, or outputSchemaRef will throw an error.
-   *
-   * The outputSchemaRef requirement ensures that Structured Output can be obtained,
-   * which is necessary for StepGate to interpret intents. Without a schema, the
-   * Flow loop cannot advance properly.
-   */
-  private validateFlowSteps(
-    stepsRegistry: ExtendedStepsRegistry,
-    logger: import("../src_common/logger.ts").Logger,
-  ): void {
-    const missingGate: string[] = [];
-    const missingTransitions: string[] = [];
-    const missingOutputSchema: string[] = [];
-
-    for (const [stepId, stepDef] of Object.entries(stepsRegistry.steps)) {
-      // Skip template sections (section.* prefix)
-      if (stepId.startsWith("section.")) {
-        continue;
-      }
-
-      const step = stepDef as PromptStepDefinition;
-
-      if (!step.structuredGate) {
-        missingGate.push(stepId);
-      }
-
-      if (!step.transitions) {
-        missingTransitions.push(stepId);
-      }
-
-      if (!step.outputSchemaRef) {
-        missingOutputSchema.push(stepId);
-      }
-    }
-
-    if (
-      missingGate.length > 0 ||
-      missingTransitions.length > 0 ||
-      missingOutputSchema.length > 0
-    ) {
-      const errors: string[] = [];
-
-      if (missingGate.length > 0) {
-        errors.push(
-          `Steps missing structuredGate: ${missingGate.join(", ")}`,
-        );
-      }
-
-      if (missingTransitions.length > 0) {
-        errors.push(
-          `Steps missing transitions: ${missingTransitions.join(", ")}`,
-        );
-      }
-
-      if (missingOutputSchema.length > 0) {
-        errors.push(
-          `Steps missing outputSchemaRef: ${missingOutputSchema.join(", ")}`,
-        );
-      }
-
-      throw new Error(
-        `[StepFlow] Flow validation failed. All Flow steps must define ` +
-          `structuredGate, transitions, and outputSchemaRef.\n${
-            errors.join("\n")
-          }\n` +
-          `See agents/docs/design/08_step_flow_design.md for requirements.`,
-      );
-    }
-
-    logger.debug(
-      `Flow validation passed: ${
-        Object.keys(stepsRegistry.steps).length
-      } steps validated`,
-    );
-  }
-
-  /**
-   * Validate completion conditions for a step.
-   */
-  private async validateCompletionConditions(
-    stepId: string,
-    _summary: IterationSummary,
-    logger: import("../src_common/logger.ts").Logger,
-  ): Promise<CompletionValidationResult> {
-    if (!this.stepsRegistry) {
-      return { valid: true };
-    }
-
-    const stepConfig = this.stepsRegistry.completionSteps?.[stepId];
-    if (!stepConfig) {
-      return { valid: true };
-    }
-
-    // Use CompletionChain for validation (logs internally)
-    if (this.completionChain) {
-      return await this.completionChain.validate(stepId, _summary);
-    }
-
-    // Fallback path - log here since CompletionChain is not used
-    logger.info(`Validating completion for step: ${stepId}`);
-
-    // Fallback to command-based validation
-    if (
-      !this.completionValidator ||
-      !stepConfig.completionConditions?.length
-    ) {
-      return { valid: true };
-    }
-
-    const result = await this.completionValidator.validate(
-      stepConfig.completionConditions,
-    );
-
-    if (result.valid) {
-      logger.info("All completion conditions passed");
-      return { valid: true };
-    }
-
-    logger.warn(`Completion validation failed: pattern=${result.pattern}`);
-
-    if (this.retryHandler && result.pattern) {
-      const retryPrompt = await this.retryHandler.buildRetryPrompt(
-        stepConfig,
-        result,
-      );
-      return { valid: false, retryPrompt };
-    }
-
-    return {
-      valid: false,
-      retryPrompt: `Completion conditions not met: ${
-        result.error ?? result.pattern
-      }`,
-    };
-  }
-
-  /**
-   * Check if AI declared completion via structured output.
-   *
-   * Only "closing" intent from Closure Step triggers completion.
-   * See design/08_step_flow_design.md Section 3 and 7.1.
-   *
-   * Note: "complete" is accepted for backward compatibility.
-   * Note: status: "completed" is NOT a completion signal - it indicates
-   *       step completion, not workflow completion.
-   */
-  private hasAICompletionDeclaration(summary: IterationSummary): boolean {
-    if (!summary.structuredOutput) {
-      return false;
-    }
-
-    const so = summary.structuredOutput;
-
-    // Only "closing" (or legacy "complete") action triggers completion validation
-    // status: "completed" alone is NOT a completion signal per 08_step_flow_design.md
-    if (isRecord(so.next_action)) {
-      const nextAction = so.next_action as Record<string, unknown>;
-      if (nextAction.action === "closing" || nextAction.action === "complete") {
-        return true;
-      }
-    }
-
-    return false;
-  }
-
-  /**
-   * Get completion step ID based on completion type.
-   */
-  private getCompletionStepId(): string {
-    if (this.completionChain) {
-      return this.completionChain.getCompletionStepId(
-        this.definition.behavior.completionType,
-      );
-    }
-    return "closure.issue";
-  }
-
-  /**
    * Delay execution for a specified duration.
    */
   private delay(ms: number): Promise<void> {
     return new Promise((resolve) => setTimeout(resolve, ms));
-  }
-
-  /**
-   * Get step ID for a given iteration.
-   *
-   * For iteration 1: Uses entryStepMapping or entryStep from registry.
-   * Entry step must be explicitly configured - no implicit fallback is allowed.
-   *
-   * For iteration > 1: Requires currentStepId to be set by structured gate routing.
-   * Errors if no routing has occurred, enforcing the documented contract that
-   * all Flow steps must define transitions.
-   */
-  private getStepIdForIteration(iteration: number): string {
-    // For iteration > 1, require routed step ID
-    if (iteration > 1) {
-      if (!this.currentStepId) {
-        throw new Error(
-          `[StepFlow] No routed step ID for iteration ${iteration}. ` +
-            `All Flow steps must define structuredGate with transitions. ` +
-            `Check ${PATHS.STEPS_REGISTRY} for missing gate configuration.`,
-        );
-      }
-      return this.currentStepId;
-    }
-
-    // For iteration 1: Use registry-based lookup
-    const completionType = this.definition.behavior.completionType;
-
-    // Try entryStepMapping first
-    if (this.stepsRegistry?.entryStepMapping?.[completionType]) {
-      return this.stepsRegistry.entryStepMapping[completionType];
-    }
-
-    // Try generic entryStep
-    if (this.stepsRegistry?.entryStep) {
-      return this.stepsRegistry.entryStep;
-    }
-
-    // No implicit fallback - entry step must be explicitly configured
-    throw new Error(
-      `[StepFlow] No entry step configured for completionType "${completionType}". ` +
-        `Define either "entryStepMapping.${completionType}" or "entryStep" in ${PATHS.STEPS_REGISTRY}.`,
-    );
-  }
-
-  /**
-   * Try to resolve a closure step prompt with adaptation override.
-   *
-   * When the current step is a closure step and the new step prompt resolver
-   * is available, resolves the closure prompt with the appropriate adaptation
-   * (e.g., "label-only", "label-and-close") based on agent config.
-   *
-   * @returns Resolved prompt or null if not a closure step or resolver unavailable
-   */
-  private async tryClosureAdaptation(
-    stepId: string,
-    ctx: RuntimeContext,
-  ): Promise<{ content: string; source: "user" | "fallback" } | null> {
-    // Skip if no step prompt resolver or no registry
-    if (!this.stepPromptResolver || !this.stepsRegistry) {
-      return null;
-    }
-
-    // Get step definition and check if it's a closure step
-    const stepDef = (this.stepsRegistry as unknown as StepRegistry).steps?.[
-      stepId
-    ];
-    if (!stepDef) {
-      return null;
-    }
-
-    const stepKind = inferStepKind(stepDef);
-    if (stepKind !== "closure") {
-      return null;
-    }
-
-    // Determine closure action from config
-    const closureAction = this.definition.github?.defaultClosureAction;
-
-    // Only use adaptation override for non-default actions
-    const overrides = closureAction && closureAction !== "close"
-      ? { adaptation: closureAction }
-      : undefined;
-
-    try {
-      const result = await this.stepPromptResolver.resolve(
-        stepId,
-        {
-          uv: {
-            issue_number: String(this.args.issue ?? ""),
-          },
-        },
-        overrides,
-      );
-
-      ctx.logger.info(
-        `[ClosureAdaptation] Resolved closure prompt for step "${stepId}"`,
-        {
-          adaptation: closureAction ?? "close",
-          source: result.source,
-          promptPath: result.promptPath,
-        },
-      );
-
-      return {
-        content: result.content,
-        source: result.source,
-      };
-    } catch (error) {
-      ctx.logger.debug(
-        `[ClosureAdaptation] Failed to resolve closure prompt, falling back to completionHandler: ${error}`,
-      );
-      return null;
-    }
-  }
-
-  /**
-   * Load JSON Schema for a step from outputSchemaRef.
-   *
-   * Implements fail-fast behavior: tracks consecutive schema resolution failures
-   * per step and throws AgentSchemaResolutionError after 2 consecutive failures.
-   *
-   * @throws AgentSchemaResolutionError after 2 consecutive failures on the same step
-   */
-  private async loadSchemaForStep(
-    stepId: string,
-    iteration: number,
-    logger: import("../src_common/logger.ts").Logger,
-  ): Promise<Record<string, unknown> | undefined> {
-    // Reset schema failure flag at the start of each load attempt
-    this.schemaResolutionFailed = false;
-
-    if (!this.stepsRegistry) {
-      return undefined;
-    }
-
-    const stepDef = this.stepsRegistry.steps[stepId] as
-      | PromptStepDefinition
-      | undefined;
-    if (!stepDef?.outputSchemaRef) {
-      logger.debug(`No outputSchemaRef for step: ${stepId}`);
-      return undefined;
-    }
-
-    // Validate outputSchemaRef format - must be object with file and schema properties
-    const ref = stepDef.outputSchemaRef;
-    if (
-      typeof ref !== "object" ||
-      ref === null ||
-      typeof ref.file !== "string" ||
-      typeof ref.schema !== "string"
-    ) {
-      const actualValue = JSON.stringify(ref);
-      const errorMsg = `Invalid outputSchemaRef format for step "${stepId}": ` +
-        `expected object with "file" and "schema" properties, got ${actualValue}. ` +
-        `See agents/docs/builder/05_troubleshooting.md for correct format.`;
-      logger.error(`[SchemaResolution] ${errorMsg}`);
-      throw new AgentSchemaResolutionError(errorMsg, {
-        stepId,
-        schemaRef: actualValue,
-        consecutiveFailures: 1,
-        iteration,
-      });
-    }
-
-    try {
-      const schema = await this.loadSchemaFromRef(
-        ref,
-        logger,
-      );
-      // Success - reset failure counter for this step
-      this.schemaFailureCount.set(stepId, 0);
-      return schema;
-    } catch (error) {
-      if (error instanceof SchemaPointerError) {
-        // Increment failure counter for this step
-        const currentCount = this.schemaFailureCount.get(stepId) ?? 0;
-        const newCount = currentCount + 1;
-        this.schemaFailureCount.set(stepId, newCount);
-
-        const schemaRef =
-          `${stepDef.outputSchemaRef.file}#${stepDef.outputSchemaRef.schema}`;
-
-        logger.error(
-          `[SchemaResolution] Failed to resolve schema pointer ` +
-            `(failure ${newCount}/${AgentRunner.MAX_SCHEMA_FAILURES})`,
-          {
-            stepId,
-            schemaRef,
-            pointer: error.pointer,
-            file: error.file,
-          },
-        );
-
-        // Check if we've hit the consecutive failure limit
-        if (newCount >= AgentRunner.MAX_SCHEMA_FAILURES) {
-          throw new AgentSchemaResolutionError(
-            `Schema resolution failed ${newCount} consecutive times for step "${stepId}". ` +
-              `Cannot resolve pointer "${error.pointer}" in ${error.file}. ` +
-              `Flow halted to prevent infinite loop.`,
-            {
-              stepId,
-              schemaRef,
-              consecutiveFailures: newCount,
-              cause: error,
-              iteration,
-            },
-          );
-        }
-
-        // Set flag to skip StepGate for this iteration (StructuredOutputUnavailable)
-        this.schemaResolutionFailed = true;
-        logger.warn(
-          `[SchemaResolution] Marking iteration as StructuredOutputUnavailable. ` +
-            `StepGate will be skipped. Fix schema reference before next iteration.`,
-        );
-        return undefined;
-      }
-      throw error;
-    }
-  }
-
-  /**
-   * Load schema from outputSchemaRef with full $ref resolution.
-   *
-   * @throws SchemaPointerError if the schema pointer cannot be resolved
-   */
-  private async loadSchemaFromRef(
-    ref: OutputSchemaRef,
-    logger: import("../src_common/logger.ts").Logger,
-  ): Promise<Record<string, unknown> | undefined> {
-    const ctx = this.getContext();
-    const schemasBase = this.stepsRegistry?.schemasBase ??
-      `.agent/${this.definition.name}/schemas`;
-    const schemasDir = join(ctx.cwd, schemasBase);
-
-    try {
-      const resolver = new SchemaResolver(schemasDir);
-      const schema = await resolver.resolve(ref.file, ref.schema);
-
-      logger.debug(`Loaded and resolved schema: ${ref.file}#${ref.schema}`);
-      return schema;
-    } catch (error) {
-      if (error instanceof Deno.errors.NotFound) {
-        logger.debug(`Schema file not found: ${join(schemasDir, ref.file)}`);
-        return undefined;
-      }
-      // Re-throw SchemaPointerError to trigger fail-fast behavior
-      if (error instanceof SchemaPointerError) {
-        throw error;
-      }
-      logger.warn(`Failed to load schema from ${ref.file}#${ref.schema}`, {
-        error: String(error),
-      });
-      return undefined;
-    }
-  }
-
-  // ============================================================================
-  // Step Flow Orchestration
-  // ============================================================================
-
-  /**
-   * Get the step context for data passing between steps.
-   */
-  getStepContext(): StepContext | null {
-    return this.stepContext;
-  }
-
-  /**
-   * Normalize stepId in structured output to match Flow's canonical value.
-   *
-   * Flow owns the authoritative stepId. The LLM only needs to provide intent
-   * (next/repeat/jump/closing) and optional targetStepId. Since stepId is
-   * defined with "const" in the schema, it should always match the expected
-   * value. If the LLM returns a different value, we correct it rather than
-   * failing, as Flow is the single source of truth.
-   *
-   * This ensures routing decisions aren't influenced by arbitrary LLM strings.
-   *
-   * @see agents/docs/design/08_step_flow_design.md
-   */
-  private normalizeStructuredOutputStepId(
-    expectedStepId: string,
-    summary: IterationSummary,
-    ctx: RuntimeContext,
-  ): void {
-    // Skip normalization if no structured output
-    if (!summary.structuredOutput) {
-      return;
-    }
-
-    const structuredOutput = summary.structuredOutput;
-
-    // Check if stepId exists in structured output
-    if (!isRecord(structuredOutput) || !isString(structuredOutput.stepId)) {
-      ctx.logger.debug(
-        `[StepFlow] No stepId in structured output, skipping normalization`,
-      );
-      return;
-    }
-
-    const actualStepId = structuredOutput.stepId;
-
-    // Normalize stepId if it differs from expected
-    if (actualStepId !== expectedStepId) {
-      // Get step name for telemetry
-      const stepName = this.stepsRegistry?.steps[expectedStepId]?.name ??
-        expectedStepId;
-      // Get issue number if available
-      const issueNumber = this.args.issue;
-
-      ctx.logger.warn(
-        `[StepFlow] stepId corrected: "${actualStepId}" -> "${expectedStepId}" ` +
-          `(Flow owns canonical stepId)`,
-        {
-          step: stepName,
-          expectedStepId,
-          actualStepId,
-          ...(issueNumber !== undefined && { issue: issueNumber }),
-        },
-      );
-
-      // Correct the value (Flow is single source of truth)
-      (structuredOutput as Record<string, unknown>).stepId = expectedStepId;
-    } else {
-      ctx.logger.debug(
-        `[StepFlow] stepId matches expected: ${actualStepId}`,
-      );
-    }
-  }
-
-  /**
-   * Record step output to step context.
-   */
-  private recordStepOutput(stepId: string, summary: IterationSummary): void {
-    if (!this.stepContext) return;
-
-    const output: Record<string, unknown> = {};
-
-    if (summary.structuredOutput) {
-      Object.assign(output, summary.structuredOutput);
-    }
-
-    output.iteration = summary.iteration;
-    output.sessionId = summary.sessionId;
-
-    if (summary.errors.length > 0) {
-      output.hasErrors = true;
-      output.errorCount = summary.errors.length;
-    }
-
-    this.stepContext.set(stepId, output);
-    this.currentStepId = stepId;
-
-    this.getContext().logger.debug(
-      `[StepFlow] Recorded output for step: ${stepId}`,
-      { outputKeys: Object.keys(output) },
-    );
-  }
-
-  /**
-   * Handle step transition using structured gate routing.
-   *
-   * All Flow steps must define structuredGate with transitions.
-   * Returns null if prerequisites are not met (no registry, no structured output,
-   * or step not found), which will cause getStepIdForIteration() to error
-   * on the next iteration.
-   *
-   * When schemaResolutionFailed is true, StepGate routing is skipped because
-   * structured output is unavailable. The step will be retried on the next
-   * iteration with the same stepId.
-   */
-  private handleStepTransition(
-    stepId: string,
-    summary: IterationSummary,
-    ctx: RuntimeContext,
-  ): RoutingResult | null {
-    // Skip StepGate when schema resolution failed (StructuredOutputUnavailable)
-    if (this.schemaResolutionFailed) {
-      ctx.logger.info(
-        `[StepFlow] Skipping StepGate routing: StructuredOutputUnavailable for step "${stepId}"`,
-      );
-      // Keep currentStepId unchanged to retry the same step
-      return null;
-    }
-    return this.tryStructuredGateRouting(stepId, summary, ctx);
-  }
-
-  /**
-   * Try to route using structured gate if configured.
-   */
-  private tryStructuredGateRouting(
-    stepId: string,
-    summary: IterationSummary,
-    ctx: RuntimeContext,
-  ): RoutingResult | null {
-    // Check prerequisites
-    if (
-      !this.stepsRegistry || !this.stepGateInterpreter || !this.workflowRouter
-    ) {
-      return null;
-    }
-
-    if (!summary.structuredOutput) {
-      return null;
-    }
-
-    // Get step definition
-    const stepDef = this.stepsRegistry.steps[stepId] as
-      | PromptStepDefinition
-      | undefined;
-    if (!stepDef?.structuredGate) {
-      return null;
-    }
-
-    // Get step kind for logging
-    const stepKind = inferStepKind(stepDef);
-
-    // Interpret structured output through the gate
-    const interpretation = this.stepGateInterpreter.interpret(
-      summary.structuredOutput,
-      stepDef,
-    );
-
-    ctx.logger.info(`[StepFlow] Interpreted intent: ${interpretation.intent}`, {
-      stepId,
-      stepKind,
-      target: interpretation.target,
-      usedFallback: interpretation.usedFallback,
-      reason: interpretation.reason,
-    });
-
-    // Merge handoff into step context
-    if (interpretation.handoff && this.stepContext) {
-      this.stepContext.set(stepId, interpretation.handoff);
-      ctx.logger.debug(`[StepFlow] Stored handoff data for step: ${stepId}`, {
-        handoffKeys: Object.keys(interpretation.handoff),
-      });
-    }
-
-    // Route to next step
-    const routing = this.workflowRouter.route(stepId, interpretation);
-
-    // Log warning if present (e.g., handoff from initial step)
-    if (routing.warning) {
-      ctx.logger.warn(`[StepFlow] ${routing.warning}`);
-    }
-
-    ctx.logger.info(
-      `[StepFlow] Routing decision: ${stepId} -> ${routing.nextStepId}`,
-      {
-        stepKind,
-        intent: interpretation.intent,
-        signalCompletion: routing.signalCompletion,
-        reason: routing.reason,
-      },
-    );
-
-    // Update current step ID
-    if (!routing.signalCompletion && routing.nextStepId !== stepId) {
-      this.currentStepId = routing.nextStepId;
-    }
-
-    return routing;
-  }
-
-  /**
-   * Check if flow routing is enabled (StepGateInterpreter and WorkflowRouter initialized).
-   * This is used for R4 fail-fast check to only enforce intent requirement when flow routing is active.
-   */
-  private hasFlowRoutingEnabled(): boolean {
-    return this.stepGateInterpreter !== null && this.workflowRouter !== null;
-  }
-
-  // ============================================================================
-  // Boundary Hook
-  // ============================================================================
-
-  /**
-   * Invoke boundary hook when a closure step emits `closing` intent.
-   *
-   * Boundary hook is the single surface for external side effects:
-   * - Issue close
-   * - Release publish
-   * - PR merge
-   *
-   * Work/Verification steps cannot mutate issues directly - all external
-   * state changes must flow through this hook.
-   *
-   * @see agents/docs/design/08_step_flow_design.md Section 7.1
-   */
-  private async invokeBoundaryHook(
-    stepId: string,
-    summary: IterationSummary,
-    ctx: RuntimeContext,
-  ): Promise<void> {
-    // Only invoke for closure steps
-    const stepDef = this.stepsRegistry?.steps[stepId] as
-      | PromptStepDefinition
-      | undefined;
-    const stepKind = stepDef ? inferStepKind(stepDef) : undefined;
-
-    if (stepKind !== "closure") {
-      ctx.logger.debug(
-        `[BoundaryHook] Skipping: step "${stepId}" is not a closure step (kind: ${
-          stepKind ?? "unknown"
-        })`,
-      );
-      return;
-    }
-
-    // Emit boundaryHook event for external handlers
-    await this.eventEmitter.emit("boundaryHook", {
-      stepId,
-      stepKind,
-      structuredOutput: summary.structuredOutput,
-    });
-
-    ctx.logger.info(`[BoundaryHook] Invoked for closure step: ${stepId}`);
-
-    // Delegate to completionHandler for actual side effects
-    // The completionHandler is responsible for implementing the boundary actions
-    // based on the agent's configuration
-    if (ctx.completionHandler.onBoundaryHook) {
-      await ctx.completionHandler.onBoundaryHook({
-        stepId,
-        stepKind,
-        structuredOutput: summary.structuredOutput,
-      });
-    }
-  }
-
-  // ============================================================================
-  // PreToolUse Hook Factory
-  // ============================================================================
-
-  /**
-   * Create a PreToolUse hook callback that blocks boundary bash commands.
-   *
-   * This hook is used to enforce the policy that Work/Verification steps
-   * cannot execute boundary actions like `gh issue close`, `gh pr merge`, etc.
-   *
-   * @param stepKind - Current step kind (work, verification, or closure)
-   * @param ctx - Runtime context for logging
-   * @returns Hook callback function for SDK PreToolUse event
-   *
-   * @see agents/docs/design/08_step_flow_design.md Section 2.1
-   * @see agents/common/tool-policy.ts
-   */
-  private createBoundaryBashBlockingHook(
-    stepKind: StepKind,
-    ctx: RuntimeContext,
-  ): (
-    input: { tool_name: string; tool_input: Record<string, unknown> },
-    toolUseId: string | undefined,
-    options: { signal: AbortSignal },
-  ) => Promise<Record<string, unknown>> {
-    return (input, _toolUseId, _options) => {
-      // Only check Bash commands
-      if (input.tool_name !== "Bash") {
-        return Promise.resolve({});
-      }
-
-      const command = input.tool_input.command as string | undefined;
-      if (!command) {
-        return Promise.resolve({});
-      }
-
-      // Check if command is allowed for this step kind
-      const result = isBashCommandAllowed(command, stepKind);
-
-      if (!result.allowed) {
-        ctx.logger.warn(
-          `[ToolPolicy] Boundary bash command blocked in ${stepKind} step`,
-          {
-            command: command.substring(0, TRUNCATION.BASH_COMMAND),
-            reason: result.reason,
-          },
-        );
-
-        return Promise.resolve({
-          hookSpecificOutput: {
-            hookEventName: "PreToolUse",
-            permissionDecision: "deny",
-            permissionDecisionReason: result.reason,
-          },
-        });
-      }
-
-      return Promise.resolve({});
-    };
   }
 }
